add READ_MEDIA_AUDIO permission and refactor navigation imports
