+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,8 @@
 
     <uses-permission android:name="android.permission.CAMERA" />
 
+    <uses-permission android:name="android.permission.READ_MEDIA_AUDIO" />
+
     <application
         android:name=".NyansapoTeaching"
         android:allowBackup="true"
