+++ app/src/main/java/com/nyansapoai/teaching/presentation/assessments/IndividualAssessment/IndividualAssessment.kt
@@ -5,7 +5,6 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -28,18 +27,16 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.nyansapoai.teaching.navController
 import com.nyansapoai.teaching.presentation.camps.LearningLevelDescription
 import com.nyansapoai.teaching.presentation.camps.components.LearningLevelItem
 import com.nyansapoai.teaching.presentation.common.components.AppCircularLoading
-import com.nyansapoai.teaching.presentation.navigation.ConductAssessmentPage
+import com.nyansapoai.teaching.navigation.ConductAssessmentPage
 import com.nyansapoai.teaching.utils.ResultStatus
 import org.koin.androidx.compose.koinViewModel
