+++ app/src/main/java/com/nyansapoai/teaching/presentation/authentication/otp/OTP.kt
@@ -20,7 +20,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -34,9 +33,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.nyansapoai.teaching.R
 import com.nyansapoai.teaching.navController
 import com.nyansapoai.teaching.presentation.authentication.otp.components.OTPImplementation
-import com.nyansapoai.teaching.presentation.authentication.otp.components.PhoneAuth
 import com.nyansapoai.teaching.presentation.common.components.CodeTextField
-import com.nyansapoai.teaching.presentation.navigation.OnboardingPage
+import com.nyansapoai.teaching.navigation.OnboardingPage
 import com.nyansapoai.teaching.utils.Utils
 import org.koin.androidx.compose.koinViewModel
